% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{AusTraits Data Portal Server Logic}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input list generated by Shiny}

\item{output}{Output list generated by Shiny}

\item{session}{Session id for Shiny interaction}
}
\description{
AusTraits Data Portal Server Logic
}
