% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{format_database_for_display}
\alias{format_database_for_display}
\title{Format relational database for display}
\usage{
format_database_for_display(database)
}
\arguments{
\item{database}{traits.build object}
}
\description{
Format relational database for display
}
