% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{format_database_for_download}
\alias{format_database_for_download}
\title{Format database for download handler}
\usage{
format_database_for_download(database)
}
\arguments{
\item{database}{traits.build object}
}
\description{
Format database for download handler
}
