% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_supabase.R
\name{init_supabase_telemetry}
\alias{init_supabase_telemetry}
\title{Initialize Supabase Telemetry}
\usage{
init_supabase_telemetry(url, key)
}
\arguments{
\item{url}{Supabase project URL.}

\item{key}{Supabase service or anon API key.}
}
\description{
Initialize Supabase Telemetry
}
