% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_supabase.R
\name{start_telemetry_session}
\alias{start_telemetry_session}
\title{Start telemetry session (works in both cloud and local mode)}
\usage{
start_telemetry_session()
}
\description{
Start telemetry session (works in both cloud and local mode)
}
