% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-server.R
\name{austraits_server}
\alias{austraits_server}
\title{AusTraits Data Portal Server Logic}
\usage{
austraits_server(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
AusTraits Data Portal Server Logic
}
