% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_supabase.R
\name{read_telemetry_metrics}
\alias{read_telemetry_metrics}
\title{Read telemetry metrics (works in both cloud and local mode)}
\usage{
read_telemetry_metrics(from_date = "2020-01-01", to_date = NULL)
}
\description{
Read telemetry metrics (works in both cloud and local mode)
}
